+++ packages/common/test/sites/_internal/SiteUiSchemaFollowers.test.ts
@@ -69,7 +69,8 @@ describe("buildUiSchema: site followers", () => {
               scope: "/properties/_followers/properties/isDiscussable",
               type: "Control",
               options: {
-                control: "hub-field-input-radio",
+                control: "hub-field-input-tile-select",
+                layout: "horizontal",
                 labels: [
                   `{{some.scope.fields.followers.discussable.enabled.label:translate}}`,
                   `{{some.scope.fields.followers.discussable.disabled.label:translate}}`,
@@ -79,6 +80,8 @@ describe("buildUiSchema: site followers", () => {
                   `{{some.scope.fields.followers.discussable.disabled.description:translate}}`,
                 ],
                 icons: ["speech-bubbles", "circle-disallowed"],
+                type: "radio",
+                styles: { "max-width": "45rem" },
               },
             },
           ],